Updated SQLite storage from separate project to keeperSDK
