+++ KeeperSdk/KeeperSdk.csproj
@@ -28,6 +28,7 @@
 
   <ItemGroup>
     <PackageReference Include="Google.Protobuf" Version="3.28.3" />
+    <PackageReference Include="Microsoft.Data.Sqlite" Version="9.0.2" />
     <PackageReference Include="Portable.BouncyCastle" Version="1.9.0" Condition="'$(TargetFramework)' == 'netstandard2.0'" />
   </ItemGroup>
