--- PowerCommanderStorageUtils/PowerCommanderStorageUtils.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <LangVersion>10</LangVersion>
-    <RootNamespace>PowerCommanderStorageUtils</RootNamespace>
-    <AssemblyName>PowerCommanderStorageUtils</AssemblyName>
-    <Description>SQLite vault storage helper for PowerCommander. Use when -UseOfflineStorage is set.</Description>
-    <GeneratePackageOnBuild>false</GeneratePackageOnBuild>
-    <CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
-  </PropertyGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\KeeperSdk\KeeperSdk.csproj" />
-    <PackageReference Include="Microsoft.Data.Sqlite" Version="9.0.2" />
-  </ItemGroup>
-</Project>
